recoani_batch/calc/requirements.txt
Django>=2.0,<3.0
mysqlclient
mysql-connector-python
uwsgi
requests
numpy
gensim
mecab-python3
tqdm
pymagnitude
